// src/app/models/backend/career-open-position.model.ts
export class CareerOpenPosition {
    id?: any;
    job_title?: String;
    job_location?: String;
    type_of_work_contract?: String;
    group_position?: String;
    image_file?: String;
    about_kaigangames?: any;
    job_role?: any;
    job_responsibilities?: any;
    job_requirement?: any;
    job_requirement_bonus_point?: any;
    job_benefits_and_perks?: any;
    current_project?: any;
    footer?: any;
    createdAt?: any;
}